add test mode
