--- src/features/ControllerFactory.ts
@@ -1,3 +1,0 @@
-export class ControllerFactory {
-  static createTaskController() {}
-}
