+++ src/features/constants.ts
@@ -1,3 +1,6 @@
 // API
 export const BASE_URL = "http://127.0.0.1:8000/api";
 export const TIME_OUT = 5000;
+
+// test
+export const IS_TEST_MODE = true;
